Added a basic particle filter.
See the ParticleFilterTests.py file for an example.
The blue dot is the "truth" and the colormap represents a gradient that
is being measured. The particles don't actually track the dot very well
right now because the observation function isn't tuned and the
measurement gradient is pretty flat, but clearly the process is more or
less working.