Stupid hack to enable showing laser in Rviz
So until now, we could not display the simulated laser returns from
gazebo in rviz. As we only want each lidar to have 1 return, we set the
number of samples to 1. Unfortunately, because of how the gazebo laser
plugin seems to works, this sets the angular resolution in the sensor
message to either nan or inf (depending on exactly how you try to tweak
it). Rviz does not know how to graph these messages. Therefore, in order
to get rviz working, we have to return 2 separate ranges for each lidar.
The two returns are in the same direction (within .00001 rad), but
because the sensor model adds noise, the returns will not be identical.
I think as long as all of our code always extracts the first range
element from the laser message, then nothing bad will ever happen (i.e.
the dummy range will never affect anything).

The .rviz schema has also been updated to display the laser ranges.